backend/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.0.6
django-cors-headers==3.14.0
django-countries==7.4.2
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.2.0
pytz==2022.6
requests==2.28.2
sqlparse==0.4.2
stripe==5.2.0
typing_extensions==4.4.0
urllib3==1.26.15
